src/requirements.txt
amqp==2.5.0
anyjson==0.3.3
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.41.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-timezone-field==3.0
djangorestframework==3.9.4
idna==2.8
kombu==4.6.3
numpy==1.16.4
oauthlib==3.1.0
opencv-python==4.1.0.25
Pillow==6.0.0
psycopg2==2.8.4
pycparser==2.20
PyJWT==1.7.1
python-crontab==2.3.8
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
virtualenv==16.6.0
